Add Windows Services support
